refactor(runbotics-desktop): :recycle: Refactor after CR
